pyproject.toml
[project]
name = "pfeed"
version = "0.0.5"
description = "Data pipeline for algo-trading, getting and storing both real-time and historical data made easy."
license = "Apache-2.0"
authors = [
    {name = "Stephen Yau", email = "softwareentrepreneer+pfeed@gmail.com"}
]
readme = "README.md"
keywords = ["trading", "algo-trading", "data pipeline", "ETL", "data lake", "data warehouse", "data integration", "historical data", "live data", "data streaming", "data engineering"]
classifiers = [
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Information Technology",
    "Topic :: Office/Business :: Financial",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.11"
dependencies = [
    "pfund>=0.0.2",
    "pandas>=2.3.3",
    "polars>=1.37.1",
    "pyarrow>=22.0.0",
    "pandera>=0.28.1",
    "msgspec>=0.20.0",
    "pydantic>=2.12.5",
    "narwhals>=2.15.0",
    "zstandard>=0.25.0", # REVIEW: this is included in python 3.14
    "beautifulsoup4>=4.14.3",
    "cloudpathlib[s3,gs,azure]>=0.23.0",
    "ray>=2.53.0",
    "pyzmq>=27.1.0",
    "deltalake>=1.3.2",
]

[project.optional-dependencies]
marimo = ["pfund-plot[marimo]>=0.0.1.dev3"]
jupyter = ["pfund-plot[jupyter]>=0.0.1.dev3"]
fmp = ["fmp-api-client>=0.0.1"]
yfinance = ["yfinance>=1.0"]
prefect = ["prefect>=3.6.11"]
# polars-cloud = ["polars-cloud>=0.4.2"]
dask = ["dask[complete,dataframe]>=2026.1.1"]
coiled = ["coiled>=1.130.2"]
spark = ["pyspark[connect,pandas-on-spark,sql]>=4.1.1"]
duckdb = ["duckdb>=1.4.3"]
lancedb = ["lancedb>=0.26.1"]
postgresql = ["psycopg[binary,pool]>=3.3.2"]
# "databento>=0.63.0",
# vortex = ["vortex-data>=0.55.0"]
# "minio>=7.2.20",
# kafka = ["confluent-kafka>=2.8.2"]
# polygon = ["polygon-api-client>=1.14.6"]

[project.scripts]
pfeed = "pfeed.__main__:run_cli"

[project.entry-points."pfund_shell.commands"]
pfeed = "pfeed.cli:pfeed_group"

[project.urls]
homepage = "https://pfund.ai"
repository = "https://github.com/PFund-Software-Ltd/pfeed"
documentation = "https://pfeed-docs.pfund.ai"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "pfeed"
module-root = ""

[tool.uv.sources]
pfund = { path = "../pfund", editable = true }
pfund-kit = { path = "../pfund-kit", editable = true }
fmp-api-client = { path = "../fmp-api-client", editable = true }

# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"

# [tool.hatch.build.targets.sdist]
# include = ['/pfeed', 'logging.yml', 'compose.yml']

# [tool.hatch.build.targets.wheel]
# include = ['/pfeed', 'logging.yml', 'compose.yml']
